Fix profile plugin directory lookup logic for TensorBoard's root run `.`
This fixes profile plugin logic for computing directory paths to avoid emitting an extraneous `.` path component, which isn't supported for non-local filesystems like GCS. Previously, this happened in some profile plugin code paths when encountering the TensorBoard root run, which is represented in TensorBoard's APIs and UI as a single period (`.`).

To fix this cleanly, this change refactors several bits of profile plugin directory handling to use a common utility for mapping from a TensorBoard logdir and run name to the corresponding subdirectory. The logic is just `logdir if run == '.' else os.path.join(logdir, run)` but a utility can help ensure consistent use.

I've also streamlined the logic in generate_runs() and generate_run_to_tools() so that the root run is handled consistently by just adding it to the result of `list_runs()` before we process them into directories. Previously, there was separate logic to force the inclusion of the `.` run if it wasn't actually a TensorBoard run (i.e. if the root logdir didn't contain tfevents data), and only that codepath correctly set the run directory to the bare logdir (without a `.` component), which explains the behavior seen in the bug report where deleting the tfevents file resolved the issue.

Tested manually by verifying that this fixes the issue as reproduced the bug. Modified unit test so it fails if PluginDirectory() is invoked with a non-normalized logdir path as a regression test.

PiperOrigin-RevId: 469577393