Track HBM bytes accessed by each op as a % of total bytes accessed - server side change.
Followup change will add some frontend changes to surface this info to the UI.
This can be useful to show which op category transfers the most HBM bytes.
This is similar to what we do for the time%.

For example, async copies typically show up as ~0% of the time, but they could be taking up a significant portion of the bytes.

PiperOrigin-RevId: 503590706