Enable web UI without requiring tensorboard
Functionally, tensorboard_plugin_profile has two parts: a web frontend, and a serverless python backend. The frontend is served by tensorboard, and tensorboard calls sends requests to the backend via the `get_plugin_apps` function to get a map of routes to functions.

Tensorboard has two main functions: identifying runs, and figuring out how to route traffic. Identifying runs requires a bit of work, as Tensorboard relies on some heavier functions in order to read from remote sources. Routing traffic is interesting, the way that it uses params is less than ideal. I'm not sure where the boundary for what's required for Tensorboard routing param parsing, vs our legacy parsing, but this PR aims to touch the minimal surface.

The Tensorboard UI generates an iframe with the web frontend. Rather than rely on that, we serve it ourselves. I've gone with cheroot as it's lightweight, but capable of serving production service, and it is already used in other projects.

For file handling, I've gone with fsspec, as it's extendable for different filesystems (local/remote). I originally looked into using univesal_pathlib which is built on top of fsspec, but encountered some difficulties. This should still allow users to `pip install fsspec[gcs,s3]` and be able to use gcs or s3 URLs.

For this to be as minimally invasive as possible to maintain tensorboard support, I decided to mock out all of the core tensorboard functions/classes that we call with minimal functionality. Other parts have 1:1 code pulled out to minimize risk. Ideally at some point an abstraction layer is built to where we no longer need to use this shim, but I'd like to see more usage of the individual `xprof` command before spending time on it.

PiperOrigin-RevId: 739320838